ррр

+++ Bonfire/Models/SeedsFromViewModel.cs
@@ -9,10 +9,10 @@ public class SeedsFromViewModel
     public string? Sort { get; set; }
     public string? Producer { get; set; }
     public DateTime ExpirationDate { get; set; }
-    public int? WeightPack { get; set; }
-    public int? QuantityPack { get; set; }
-    public int? AmountSeedsWeight { get; set; }
-    public int? AmountSeedsQuantity { get; set; }
+    public double? WeightPack { get; set; }
+    public double? QuantityPack { get; set; }
+    public double? AmountSeedsWeight { get; set; }
+    public double? AmountSeedsQuantity { get; set; }
 
     public bool IsStillGood => ExpirationDate.Year >= DateTime.Now.Year + 2;
     public bool IsOld => ExpirationDate.Year >= DateTime.Now.Year + 1;
